refactor(setup): split god-file + fix the post-upgrade setup that never ran
infrastructure/setup.ts (892 lines) split by responsibility:
codex-skill.ts (skill install + p. router verification - its consumers
sync-service/doctor-service/commands-setup now import it directly) and
statusline-installer.ts (script + modules + symlink + settings wiring).
setup.ts keeps the run() orchestrator + provider installers (~470).

REAL BUG FIXED on the way: bin/prjct.ts called setup.run() through a
default export that PR #132 (2026-02) deleted - bin/ sits outside core's
typecheck, so the post-upgrade re-setup threw TypeError into its catch
and silently only stamped the version, every upgrade, for 4 months.
run() is now a named export and bin imports it directly.
