+++ core/infrastructure/setup.ts
@@ -29,8 +29,7 @@ import { prjctDb } from '../storage/database'
 import { getErrorMessage, isNotFoundError } from '../types/fs'
 import type { AIProviderConfig, AIProviderName } from '../types/provider'
 import { getTimeout } from '../utils/constants'
-import { fileExists, readJson, writeJson } from '../utils/file-helper'
-import { sha256 } from '../utils/hash'
+import { fileExists } from '../utils/file-helper'
 import log from '../utils/logger'
 import { PACKAGE_ROOT, VERSION } from '../utils/version'
 import {
@@ -40,10 +39,12 @@ import {
   Providers,
   selectProvider,
 } from './ai-provider'
+import { installCodexSkill, verifyCodexPRouterReady } from './codex-skill'
 import installer from './command-installer'
 import editorsConfig from './editors-config'
 import { mergeWithMarkers } from './ide-project-installer'
 import pathManager from './path-manager'
+import { installStatusLine } from './statusline-installer'
 
 interface ProviderSetupResult {
   provider: AIProviderName
@@ -123,7 +124,7 @@ async function installAICLI(provider: AIProviderConfig): Promise<boolean> {
 /**
  * Main setup function - installs for ALL detected providers
  */
-async function run(): Promise<SetupResults> {
+export async function run(): Promise<SetupResults> {
   // Step 0: Detect all available providers
   const detection = await detectAllProviders()
   const selection = await selectProvider()
@@ -377,200 +378,6 @@ async function installAntigravitySkill(): Promise<{
   }
 }
 
-// Codex Installation (Skills-based)
-
-const CODEX_SKILL_META_MARKER = 'prjct-codex-router'
-
-import type { CodexPRouterStatus } from '../types/infrastructure.js'
-
-function getCodexSkillPath(): string {
-  return path.join(os.homedir(), '.codex', 'skills', 'prjct', 'SKILL.md')
-}
-
-function getCodexSkillMetadata(templateHash: string): string {
-  return `<!-- ${CODEX_SKILL_META_MARKER}: ${JSON.stringify({
-    version: VERSION,
-    templateHash,
-  })} -->`
-}
-
-function parseCodexSkillMetadata(
-  content: string
-): { version?: string; templateHash?: string } | null {
-  const match = content.match(
-    new RegExp(`<!--\\s*${CODEX_SKILL_META_MARKER}:\\s*(\\{[\\s\\S]*?\\})\\s*-->`)
-  )
-  if (!match) return null
-  try {
-    return JSON.parse(match[1]) as { version?: string; templateHash?: string }
-  } catch {
-    return null
-  }
-}
-
-function hashContent(content: string): string {
-  return sha256(content)
-}
-
-async function loadCodexSkillTemplate(): Promise<string | null> {
-  const bundled = getTemplateContent('codex/SKILL.md')
-  if (bundled) return bundled
-
-  const templatePath = path.join(PACKAGE_ROOT, 'templates', 'codex', 'SKILL.md')
-  if (!(await fileExists(templatePath))) {
-    return null
-  }
-
-  return fs.readFile(templatePath, 'utf-8')
-}
-
-function buildCodexSkillContent(templateContent: string): {
-  content: string
-  templateHash: string
-} {
-  const normalized = templateContent.trimEnd()
-  const templateHash = hashContent(normalized)
-  const metadata = getCodexSkillMetadata(templateHash)
-  return {
-    content: `${normalized}\n\n${metadata}\n`,
-    templateHash,
-  }
-}
-
-/**
- * Install prjct as a skill for OpenAI Codex
- *
- * Codex uses SKILL.md files in ~/.codex/skills/
- * Following the same pattern as Antigravity.
- */
-export async function installCodexSkill(): Promise<{
-  success: boolean
-  action: string | null
-}> {
-  try {
-    const skillMdPath = getCodexSkillPath()
-    const prjctSkillDir = path.dirname(skillMdPath)
-    // Ensure skills directory exists
-    await fs.mkdir(prjctSkillDir, { recursive: true })
-
-    // Check if SKILL.md already exists
-    const skillExists = await fileExists(skillMdPath)
-
-    // Read template content - try bundle first
-    const templateContent = await loadCodexSkillTemplate()
-    if (!templateContent) {
-      log.warn('Codex SKILL.md template not found')
-      return { success: false, action: null }
-    }
-
-    const built = buildCodexSkillContent(templateContent)
-
-    if (skillExists) {
-      const existing = await fs.readFile(skillMdPath, 'utf-8').catch(() => '')
-      if (existing === built.content) {
-        return { success: true, action: 'unchanged' }
-      }
-    }
-
-    // Write SKILL.md
-    await fs.writeFile(skillMdPath, built.content, 'utf-8')
-
-    return { success: true, action: skillExists ? 'updated' : 'created' }
-  } catch (error) {
-    log.warn(`Codex skill warning: ${getErrorMessage(error)}`)
-    return { success: false, action: null }
-  }
-}
-
-export async function verifyCodexPRouterReady(
-  options: { autoRepair?: boolean } = {}
-): Promise<CodexPRouterStatus> {
-  const skillPath = getCodexSkillPath()
-  const codexDetection = await detectCodex()
-  if (!codexDetection.installed) {
-    return {
-      installed: false,
-      verified: true,
-      skillPath,
-      message: 'Codex not detected',
-    }
-  }
-
-  const templateContent = await loadCodexSkillTemplate()
-  if (!templateContent) {
-    return {
-      installed: true,
-      verified: false,
-      skillPath,
-      message: 'Codex SKILL.md template missing from prjct installation',
-      fix: ['Reinstall prjct-cli package', 'Run `prjct setup`'],
-    }
-  }
-
-  const expected = buildCodexSkillContent(templateContent)
-
-  const maybeRepair = async (): Promise<boolean> => {
-    if (!options.autoRepair) return false
-    const result = await installCodexSkill()
-    return result.success
-  }
-
-  let skillContent = ''
-  if (!(await fileExists(skillPath))) {
-    if (!(await maybeRepair())) {
-      return {
-        installed: true,
-        verified: false,
-        skillPath,
-        templateHash: expected.templateHash,
-        message: 'Codex skill missing at ~/.codex/skills/prjct/SKILL.md',
-        fix: ['Run `prjct start` to install Codex skill'],
-      }
-    }
-  }
-
-  skillContent = await fs.readFile(skillPath, 'utf-8').catch(() => '')
-  let metadata = parseCodexSkillMetadata(skillContent)
-  const metadataMatches =
-    metadata?.version === VERSION && metadata?.templateHash === expected.templateHash
-
-  if (!metadataMatches) {
-    if (!(await maybeRepair())) {
-      return {
-        installed: true,
-        verified: false,
-        skillPath,
-        templateHash: expected.templateHash,
-        message: 'Codex skill metadata mismatch (outdated router)',
-        fix: ['Run `prjct start` or `prjct setup` to refresh Codex skill'],
-      }
-    }
-
-    skillContent = await fs.readFile(skillPath, 'utf-8').catch(() => '')
-    metadata = parseCodexSkillMetadata(skillContent)
-    const repaired =
-      metadata?.version === VERSION && metadata?.templateHash === expected.templateHash
-    if (!repaired) {
-      return {
-        installed: true,
-        verified: false,
-        skillPath,
-        templateHash: expected.templateHash,
-        message: 'Codex skill could not be repaired automatically',
-        fix: ['Delete ~/.codex/skills/prjct/SKILL.md', 'Run `prjct setup`'],
-      }
-    }
-  }
-
-  return {
-    installed: true,
-    verified: true,
-    skillPath,
-    templateHash: expected.templateHash,
-    message: 'Codex p. router ready',
-  }
-}
-
 /**
  * Migrate existing projects to add cliVersion field
  * This clears the status line warning after npm update
@@ -615,235 +422,6 @@ async function migrateProjectsCliVersion(): Promise<void> {
     if (!isNotFoundError(error)) {
       // Log unexpected errors but don't crash - migration is optional
       log.warn(`Migration warning: ${getErrorMessage(error)}`)
-    }
-  }
-}
-
-/**
- * Ensure settings.json has statusLine configured
- */
-async function ensureStatusLineSettings(
-  settingsPath: string,
-  statusLinePath: string
-): Promise<void> {
-  let settings: Record<string, unknown> = {}
-  if (await fileExists(settingsPath)) {
-    try {
-      settings = (await readJson<Record<string, unknown>>(settingsPath)) ?? {}
-    } catch (error) {
-      // Invalid JSON, start fresh - but propagate unexpected errors
-      if (!(error instanceof SyntaxError)) {
-        throw error
-      }
-    }
-  }
-  settings.statusLine = { type: 'command', command: statusLinePath }
-  await writeJson(settingsPath, settings)
-}
-
-/**
- * Install status line script with version check
- * Copies modular statusline from assets/ to ~/.prjct-cli/statusline/
- * Includes: statusline.sh, lib/, components/, themes/, config.json
- * Creates symlink at ~/.claude/prjct-statusline.sh
- * Updates CLI_VERSION in the script
- */
-async function installStatusLine(): Promise<void> {
-  try {
-    const claudeDir = path.join(os.homedir(), '.claude')
-    const settingsPath = path.join(claudeDir, 'settings.json')
-    const claudeStatusLinePath = path.join(claudeDir, 'prjct-statusline.sh')
-
-    // Target location for the actual script
-    const prjctStatusLineDir = pathManager.getStatusLinePath()
-    const prjctStatusLinePath = path.join(prjctStatusLineDir, 'statusline.sh')
-    const prjctThemesDir = path.join(prjctStatusLineDir, 'themes')
-    const prjctLibDir = path.join(prjctStatusLineDir, 'lib')
-    const prjctComponentsDir = path.join(prjctStatusLineDir, 'components')
-    const prjctConfigPath = path.join(prjctStatusLineDir, 'config.json')
-
-    // Source assets (from the package)
-    const assetsDir = path.join(PACKAGE_ROOT, 'assets', 'statusline')
-    const sourceScript = path.join(assetsDir, 'statusline.sh')
-    const sourceThemeDir = path.join(assetsDir, 'themes')
-    const sourceLibDir = path.join(assetsDir, 'lib')
-    const sourceComponentsDir = path.join(assetsDir, 'components')
-    const sourceConfigPath = path.join(assetsDir, 'default-config.json')
-
-    // Ensure directories exist
-    if (!(await fileExists(claudeDir))) {
-      await fs.mkdir(claudeDir, { recursive: true })
-    }
-    if (!(await fileExists(prjctStatusLineDir))) {
-      await fs.mkdir(prjctStatusLineDir, { recursive: true })
-    }
-    if (!(await fileExists(prjctThemesDir))) {
-      await fs.mkdir(prjctThemesDir, { recursive: true })
-    }
-    if (!(await fileExists(prjctLibDir))) {
-      await fs.mkdir(prjctLibDir, { recursive: true })
-    }
-    if (!(await fileExists(prjctComponentsDir))) {
-      await fs.mkdir(prjctComponentsDir, { recursive: true })
-    }
-
-    // Check if statusline already exists
-    if (await fileExists(prjctStatusLinePath)) {
-      const existingContent = await fs.readFile(prjctStatusLinePath, 'utf8')
-
-      if (existingContent.includes('CLI_VERSION=')) {
-        // Has CLI_VERSION - update if needed
-        const versionMatch = existingContent.match(/CLI_VERSION="([^"]*)"/)
-
-        if (versionMatch && versionMatch[1] !== VERSION) {
-          // Update CLI_VERSION in-place
-          const updatedContent = existingContent.replace(
-            /CLI_VERSION="[^"]*"/,
-            `CLI_VERSION="${VERSION}"`
-          )
-          await fs.writeFile(prjctStatusLinePath, updatedContent, { mode: 0o755 })
-        }
-
-        // Ensure modular structure is installed (upgrade path)
-        await installStatusLineModules(sourceLibDir, prjctLibDir)
-        await installStatusLineModules(sourceComponentsDir, prjctComponentsDir)
-
-        // Ensure symlink and settings
-        await ensureStatusLineSymlink(claudeStatusLinePath, prjctStatusLinePath)
-        await ensureStatusLineSettings(settingsPath, claudeStatusLinePath)
-        return
-      }
-      // else: Script exists WITHOUT CLI_VERSION - fall through to replace with new version
-    }
-
-    // Install fresh from assets if source exists
-    if (await fileExists(sourceScript)) {
-      // Copy script and update version
-      let scriptContent = await fs.readFile(sourceScript, 'utf8')
-      scriptContent = scriptContent.replace(/CLI_VERSION="[^"]*"/, `CLI_VERSION="${VERSION}"`)
-      await fs.writeFile(prjctStatusLinePath, scriptContent, { mode: 0o755 })
-
-      // Copy lib/ modules
-      await installStatusLineModules(sourceLibDir, prjctLibDir)
-
-      // Copy components/
-      await installStatusLineModules(sourceComponentsDir, prjctComponentsDir)
-
-      // Copy themes
-      if (await fileExists(sourceThemeDir)) {
-        const themes = await fs.readdir(sourceThemeDir)
-        for (const theme of themes) {
-          const src = path.join(sourceThemeDir, theme)
-          const dest = path.join(prjctThemesDir, theme)
-          // Always update themes to get new icons/colors
-          await fs.copyFile(src, dest)
-        }
-      }
-
-      // Copy default config (only if not exists - preserve user customizations)
-      if (!(await fileExists(prjctConfigPath)) && (await fileExists(sourceConfigPath))) {
-        await fs.copyFile(sourceConfigPath, prjctConfigPath)
-      }
-    } else {
-      // Fallback: create simple script inline
-      const scriptContent = `#!/bin/bash
-# prjct Status Line for Claude Code
-CLI_VERSION="${VERSION}"
-input=$(cat)
-CWD=$(echo "$input" | jq -r '.workspace.current_dir // "~"' 2>/dev/null)
-CONFIG="$CWD/.prjct/prjct.config.json"
-if [ -f "$CONFIG" ]; then
-  PROJECT_ID=$(jq -r '.projectId // ""' "$CONFIG" 2>/dev/null)
-  if [ -n "$PROJECT_ID" ]; then
-    PROJECT_JSON="$HOME/.prjct-cli/projects/$PROJECT_ID/project.json"
-    if [ -f "$PROJECT_JSON" ]; then
-      PROJECT_VERSION=$(jq -r '.cliVersion // ""' "$PROJECT_JSON" 2>/dev/null)
-      if [ -z "$PROJECT_VERSION" ] || [ "$PROJECT_VERSION" != "$CLI_VERSION" ]; then
-        echo "prjct v$CLI_VERSION - run p. sync"
-        exit 0
-      fi
-    else
-      echo "prjct v$CLI_VERSION - run p. sync"
-      exit 0
-    fi
-    STATE="$HOME/.prjct-cli/projects/$PROJECT_ID/storage/state.json"
-    if [ -f "$STATE" ]; then
-      TASK=$(jq -r '.currentTask.description // ""' "$STATE" 2>/dev/null)
-      if [ -n "$TASK" ]; then
-        echo "$TASK"
-        exit 0
-      fi
-    fi
-  fi
-fi
-echo "prjct"
-`
-      await fs.writeFile(prjctStatusLinePath, scriptContent, { mode: 0o755 })
-    }
-
-    // Create symlink and configure settings
-    await ensureStatusLineSymlink(claudeStatusLinePath, prjctStatusLinePath)
-    await ensureStatusLineSettings(settingsPath, claudeStatusLinePath)
-  } catch (error) {
-    // Silently fail if directories don't exist
-    if (!isNotFoundError(error)) {
-      // Log unexpected errors but don't crash - status line is optional
-      log.warn(`Status line warning: ${getErrorMessage(error)}`)
-    }
-  }
-}
-
-/**
- * Install statusline modules (lib/ or components/)
- * Copies .sh files from source to destination, always overwriting for updates
- */
-async function installStatusLineModules(sourceDir: string, destDir: string): Promise<void> {
-  if (!(await fileExists(sourceDir))) {
-    return
-  }
-
-  const files = await fs.readdir(sourceDir)
-  for (const file of files) {
-    if (file.endsWith('.sh')) {
-      const src = path.join(sourceDir, file)
-      const dest = path.join(destDir, file)
-      await fs.copyFile(src, dest)
-      await fs.chmod(dest, 0o755)
-    }
-  }
-}
-
-/**
- * Ensure symlink from Claude config to prjct statusline
- */
-async function ensureStatusLineSymlink(linkPath: string, targetPath: string): Promise<void> {
-  try {
-    // Check if link already points to correct target
-    if (await fileExists(linkPath)) {
-      const stats = await fs.lstat(linkPath)
-      if (stats.isSymbolicLink()) {
-        const existingTarget = await fs.readlink(linkPath)
-        if (existingTarget === targetPath) {
-          return // Already correct
-        }
-      }
-      // Remove existing file/symlink
-      await fs.unlink(linkPath)
-    }
-    // Create symlink
-    await fs.symlink(targetPath, linkPath)
-  } catch (_error) {
-    // If symlink fails (e.g., Windows, permission issues), try copy instead
-    try {
-      if (await fileExists(targetPath)) {
-        await fs.copyFile(targetPath, linkPath)
-        await fs.chmod(linkPath, 0o755)
-      }
-    } catch (copyError) {
-      // Both symlink and copy failed - log if unexpected error
-      if (!isNotFoundError(copyError)) {
-        log.warn(`Symlink fallback warning: ${(copyError as Error).message}`)
-      }
     }
   }
 }
