+++ core/services/doctor-service.ts
@@ -12,9 +12,9 @@
 import { execSync } from 'node:child_process'
 import path from 'node:path'
 import chalk from 'chalk'
+import { verifyCodexPRouterReady } from '../infrastructure/codex-skill'
 import configManager from '../infrastructure/config-manager'
 import pathManager from '../infrastructure/path-manager'
-import { verifyCodexPRouterReady } from '../infrastructure/setup'
 import { stateStorage } from '../storage/state-storage'
 import type { CheckResult, DoctorResult } from '../types/services/extracted'
 import { fileExists } from '../utils/file-helper'
