+++ core/services/sync-service.ts
@@ -22,10 +22,10 @@ import { indexCoChanges } from '../domain/git-cochange'
 import { indexImports } from '../domain/import-graph'
 import { getErrorMessage } from '../errors'
 import { detectCodex } from '../infrastructure/ai-provider'
+import { verifyCodexPRouterReady } from '../infrastructure/codex-skill'
 import commandInstaller from '../infrastructure/command-installer'
 import configManager from '../infrastructure/config-manager'
 import pathManager from '../infrastructure/path-manager'
-import { verifyCodexPRouterReady } from '../infrastructure/setup'
 import { analysisStorage } from '../storage/analysis-storage'
 import { ideasStorage } from '../storage/ideas-storage'
 import llmAnalysisStorage from '../storage/llm-analysis-storage'
